+++ test/context_test.exs
@@ -13,21 +13,20 @@ defmodule Sentry.ContextTest do
   end
 
   test "storing breadcrumbs appears when generating event" do
-    Sentry.Context.add_breadcrumb(%{"category" => "a category", "message" => "a message"})
+    Sentry.Context.add_breadcrumb(category: "a category", message: "a message")
     Sentry.Context.add_breadcrumb(%{"category" => "a category", "message" => "second message"})
 
     exception = RuntimeError.exception("error")
     event = Sentry.Event.transform_exception(exception, [])
 
-    first_breadcrumb = Enum.at(event.breadcrumbs, 0)
-    second_breadcrumb = Enum.at(event.breadcrumbs, 1)
+    [first_breadcrumb, second_breadcrumb] = event.breadcrumbs
 
     assert event.user == %{}
     assert event.extra == %{}
     assert event.tags == %{}
-    assert first_breadcrumb["category"] == "a category"
+    assert first_breadcrumb[:category] == "a category"
     assert second_breadcrumb["category"] == "a category"
-    assert first_breadcrumb["message"] == "a message"
+    assert first_breadcrumb[:message] == "a message"
     assert second_breadcrumb["message"] == "second message"
     assert first_breadcrumb["timestamp"] <= second_breadcrumb["timestamp"]
   end