+++ test/plug_test.exs
@@ -2,29 +2,6 @@ defmodule Sentry.PlugTest do
   use ExUnit.Case, async: true
   use Plug.Test
 
-  defmodule ExampleApp do
-    use Plug.Router
-    use Plug.ErrorHandler
-    use Sentry.Plug, request_id_header: "x-request-id"
-
-
-    plug Plug.Parsers, parsers: [:multipart]
-    plug Plug.RequestId
-    plug :match
-    plug :dispatch
-
-    get "/error_route" do
-      _ = conn
-      raise RuntimeError, "Error"
-    end
-
-    match "/error_route" do
-      _ = conn
-      raise RuntimeError, "Error"
-    end
-
-  end
-
   test "non-existent route exceptions are ignored" do
     exception = %FunctionClauseError{arity: 4,
                                      function: :do_match,
@@ -51,14 +28,14 @@ defmodule Sentry.PlugTest do
 
    assert_raise(RuntimeError, "Error", fn ->
      conn(:get, "/error_route")
-     |> ExampleApp.call([])
+     |> Sentry.ExampleApp.call([])
    end)
   end
 
   test "builds request data" do
     conn = conn(:get, "/error_route?key=value")
-    |> put_req_cookie("cookie_key", "cookie_value")
-    |> put_req_header("accept-language", "en-US")
+           |> put_req_cookie("cookie_key", "cookie_value")
+           |> put_req_header("accept-language", "en-US")
 
     request_data = Sentry.Plug.build_request_interface_data(conn, [header_scrubber: &Sentry.Plug.default_header_scrubber/1])
 
@@ -78,7 +55,7 @@ defmodule Sentry.PlugTest do
     conn = conn(:post, "/error_route", %{
       "hello" => "world",
       "password" => "test",
-      "cc" => "4242424242424242" })
+      "cc" => "4242424242424242"})
     |> put_req_cookie("cookie_key", "cookie_value")
     |> put_req_header("accept-language", "en-US")
     |> put_req_header("authorization", "ignorme")
@@ -119,8 +96,12 @@ defmodule Sentry.PlugTest do
 
     request_data = Sentry.Plug.build_request_interface_data(conn, body_scrubber: &Sentry.Plug.default_body_scrubber/1)
     assert request_data[:method] == "POST"
-    assert request_data[:data] == %{"secret" => "*********", "password" => "*********",
-      "passwd" => "*********", "credit_card" => "*********", "cc" => "*********",
+    assert request_data[:data] == %{
+      "secret" => "*********",
+      "password" => "*********",
+      "passwd" => "*********",
+      "credit_card" => "*********",
+      "cc" => "*********",
       "another_cc" => "*********"}
   end